+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.1] - 2022-03-31
+
+Some minor improvements in transaction builder.
+
+### Added
+
+- Add more details to the message of expection when UTxO selectors failed.
+- Validate output value is non-negative.
+
+
+
 ## [0.3.0] - 2022-03-21
 
 ### Added